requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
boto3==1.34.87
botocore==1.34.87
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
MarkupSafe==2.1.5
numpy==1.26.4
oauthlib==3.2.2
openai==1.14.3
pandas==2.2.1
pillow==10.3.0
praw==7.7.1
prawcore==2.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
StrEnum==0.4.15
threadpoolctl==3.4.0
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
update-checker==0.18.0
urllib3==2.2.1
websocket-client==1.7.0
Werkzeug==3.0.2
